Finishing first draft

+++ 123TruckHelper/Models/API/NotificationTruckResponse.cs
@@ -4,7 +4,7 @@
     {
         public int NotificationID { get; set; }
 
-        public double Revenue { get; set; }
+        public decimal Revenue { get; set; }
 
         public double Proft { get; set; }
 
@@ -16,7 +16,7 @@
 
         public double DestLon { get; set; }
 
-        public double TripDist { get; set; }
+        public decimal TripDist { get; set; }
 
         public double DistToStart { get; set; }
